Moved the Color and ColorSet classes from the ColorSchemeService into their own new file.
