+++ src/app/components/aggregation/subcomponent.chartjs.histogram.ts
@@ -17,7 +17,7 @@ import { ElementRef } from '@angular/core';
 import { AggregationOptions } from './aggregation.component';
 import { AggregationSubcomponentListener } from './subcomponent.aggregation.abstract';
 import { ChartJsBarSubcomponent, ChartJsBarDataset } from './subcomponent.chartjs.bar';
-import { Color } from '../../services/color-scheme.service';
+import { Color } from '../../color';
 
 export class ChartJsHistogramSubcomponent extends ChartJsBarSubcomponent {
     /**
