+++ src/app/components/aggregation/subcomponent.chartjs.pie.ts
@@ -17,7 +17,7 @@ import { ElementRef } from '@angular/core';
 import { AbstractChartJsDataset, AbstractChartJsSubcomponent, ChartJsData } from './subcomponent.chartjs.abstract';
 import { AggregationOptions } from './aggregation.component';
 import { AggregationSubcomponentListener } from './subcomponent.aggregation.abstract';
-import { Color } from '../../services/color-scheme.service';
+import { Color } from '../../color';
 
 import * as _ from 'lodash';
