+++ src/app/components/aggregation/subcomponent.list.spec.ts
@@ -18,7 +18,7 @@
 
 import { ListSubcomponent } from './subcomponent.list';
 import { AggregationSubcomponentListener, AggregationSubcomponentOptions } from './subcomponent.aggregation.abstract';
-import { Color } from '../../services/color-scheme.service';
+import { Color } from '../../color';
 import { ElementRef } from '@angular/core';
 
 class TestAggregationSubcomponentOptions implements AggregationSubcomponentOptions {
